Conflicts:
	ArduPlane/ArduPlane.pde
	ArduPlane/Attitude.pde
	ArduPlane/GCS_Mavlink.pde
	ArduPlane/Parameters.pde
	ArduPlane/release-notes.txt
	ArduPlane/system.pde
	ArduPlane/takeoff.pde
	Tools/scripts/build_vrbrain_binaries.sh
	Tools/scripts/build_vrbrain_branch_binaries.sh
	Tools/scripts/build_vrbrain_project_binaries.sh
	Tools/scripts/install-travis-env.sh
	libraries/AP_AHRS/AP_AHRS_DCM.h
	libraries/AP_Airspeed/AP_Airspeed.cpp
	libraries/AP_Arming/AP_Arming.cpp
	libraries/AP_BattMonitor/AP_BattMonitor_Analog.h
	libraries/AP_Compass/AP_Compass_VRBRAIN.cpp
	libraries/AP_Frsky_Telem/AP_Frsky_Telem.cpp
	libraries/AP_GPS/AP_GPS.cpp
	libraries/AP_GPS/AP_GPS.h
	libraries/AP_HAL_AVR_SITL/sitl_gps.cpp
	libraries/AP_HAL_SITL/RCInput.cpp
	libraries/AP_HAL_VRBRAIN/AnalogIn.cpp
	libraries/AP_HAL_VRBRAIN/AnalogIn.h
	libraries/AP_HAL_VRBRAIN/HAL_VRBRAIN_Class.cpp
	libraries/AP_HAL_VRBRAIN/Util.cpp
	libraries/AP_InertialSensor/AP_InertialSensor.cpp
	libraries/AP_InertialSensor/AP_InertialSensor.h
	libraries/AP_InertialSensor/AP_InertialSensor_PX4.cpp
	libraries/AP_Notify/AP_Notify.cpp
	libraries/AP_Notify/AP_Notify.h
	libraries/AP_Notify/Buzzer.cpp
	libraries/AP_Notify/ToneAlarm_Linux.cpp
	libraries/AP_SerialManager/AP_SerialManager.cpp
	libraries/AP_SerialManager/AP_SerialManager.h
	libraries/DataFlash/DataFlash_File.cpp
	libraries/DataFlash/LogFile.cpp
	libraries/GCS_MAVLink/GCS.h
	libraries/GCS_MAVLink/GCS_Common.cpp
	libraries/GCS_MAVLink/GCS_MAVLink.cpp
	libraries/GCS_MAVLink/GCS_MAVLink.h
	libraries/SITL/SITL.cpp
	libraries/SITL/SITL.h
	mk/VRBRAIN/ROMFS_VRBRAIN52_APM/init.d/rc.error
	mk/VRBRAIN/ROMFS_VRBRAIN52_APM/init.d/rcS
	mk/VRBRAIN/ROMFS_VRUBRAIN52_APM/init.d/rc.error
	mk/VRBRAIN/ROMFS_VRUBRAIN52_APM/init.d/rcS
	mk/VRBRAIN/config_vrbrain-v45P_APM.mk
	mk/VRBRAIN/config_vrbrain-v51P_APM.mk
	mk/VRBRAIN/config_vrbrain-v51ProP_APM.mk
	mk/VRBRAIN/config_vrbrain-v51Pro_APM.mk
	mk/VRBRAIN/config_vrbrain-v52P_APM.mk
	mk/VRBRAIN/config_vrbrain-v52ProP_APM.mk
	mk/VRBRAIN/config_vrbrain-v52Pro_APM.mk
	mk/VRBRAIN/config_vrbrain-v52_APM.mk
	mk/VRBRAIN/config_vrubrain-v51P_APM.mk
	mk/VRBRAIN/config_vrubrain-v52_APM.mk
	mk/targets.mk
	mk/vrbrain_targets.mk

+++ libraries/AP_Math/examples/location/location.pde
@@ -22,7 +22,7 @@
 #include <AP_Declination.h>
 
 #include <AP_HAL_AVR.h>
-#include <AP_HAL_AVR_SITL.h>
+#include <AP_HAL_SITL.h>
 #include <AP_HAL_Empty.h>
 #include <AP_HAL_Linux.h>
 #include <AP_AHRS.h>
@@ -32,6 +32,8 @@
 #include <AP_Vehicle.h>
 #include <AP_ADC_AnalogSource.h>
 #include <AP_Compass.h>
+#include <AP_BattMonitor.h>
+#include <AP_RangeFinder.h>
 
 const AP_HAL::HAL& hal = AP_HAL_BOARD_DRIVER;
 
@@ -39,12 +41,12 @@ static const struct {
     Vector2f wp1, wp2, location;
     bool passed;
 } test_points[] = {
-    { Vector2f(-35.3647759314918, 149.16265692810987),
-      Vector2f(-35.36279922658029, 149.16352169591426),
-      Vector2f(-35.36214956969903, 149.16461410046492), true },
-    { Vector2f(-35.36438601157189, 149.16613916088568),
-      Vector2f(-35.364432558610254, 149.16287313113048),
-      Vector2f(-35.36491510034746, 149.16365837225004), false },
+    { Vector2f(-35.3647759314918f, 149.16265692810987f),
+      Vector2f(-35.36279922658029f, 149.16352169591426f),
+      Vector2f(-35.36214956969903f, 149.16461410046492f), true },
+    { Vector2f(-35.36438601157189f, 149.16613916088568f),
+      Vector2f(-35.364432558610254f, 149.16287313113048f),
+      Vector2f(-35.36491510034746f, 149.16365837225004f), false },
     { Vector2f(0, 0),
       Vector2f(0, 1),
       Vector2f(0, 2), true },
@@ -67,8 +69,8 @@ static const struct {
 static struct Location location_from_point(Vector2f pt)
 {
     struct Location loc = {0};
-    loc.lat = pt.x * 1.0e7;
-    loc.lng = pt.y * 1.0e7;
+    loc.lat = pt.x * 1.0e7f;
+    loc.lng = pt.y * 1.0e7f;
     return loc;
 }
 
@@ -100,7 +102,7 @@ static void test_one_offset(const struct Location &loc,
     hal.console->printf("location_offset took %u usec\n",
                         (unsigned)(hal.scheduler->micros() - t1));
     dist2 = get_distance(loc, loc2);
-    bearing2 = get_bearing_cd(loc, loc2) * 0.01;
+    bearing2 = get_bearing_cd(loc, loc2) * 0.01f;
     float brg_error = bearing2-bearing;
     if (brg_error > 180) {
         brg_error -= 360;
@@ -108,8 +110,8 @@ static void test_one_offset(const struct Location &loc,
         brg_error += 360;
     }
 
-    if (fabsf(dist - dist2) > 1.0 ||
-        brg_error > 1.0) {
+    if (fabsf(dist - dist2) > 1.0f ||
+        brg_error > 1.0f) {
         hal.console->printf("Failed offset test brg_error=%f dist_error=%f\n",
                       brg_error, dist-dist2);
     }
@@ -118,8 +120,8 @@ static void test_one_offset(const struct Location &loc,
 static const struct {
     float ofs_north, ofs_east, distance, bearing;
 } test_offsets[] = {
-    { 1000, 1000,  sqrt(2.0)*1000, 45 },
-    { 1000, -1000, sqrt(2.0)*1000, -45 },
+    { 1000, 1000,  sqrt(2.0f)*1000, 45 },
+    { 1000, -1000, sqrt(2.0f)*1000, -45 },
     { 1000, 0,     1000, 0 },
     { 0, 1000,     1000, 90 },
 };
@@ -128,8 +130,8 @@ static void test_offset(void)
 {
     struct Location loc;
 
-    loc.lat = -35*1.0e7;
-    loc.lng = 149*1.0e7;
+    loc.lat = -35*1.0e7f;
+    loc.lng = 149*1.0e7f;
 
     for (uint8_t i=0; i<ARRAY_LENGTH(test_offsets); i++) {
         test_one_offset(loc,
@@ -252,7 +254,7 @@ static void test_wrap_cd(void)
 
     for (uint8_t i=0; i<sizeof(wrap_PI_tests)/sizeof(wrap_PI_tests[0]); i++) {
         float r = wrap_PI(wrap_PI_tests[i].v);
-        if (fabs(r - wrap_PI_tests[i].wv) > 0.001f) {
+        if (fabsf(r - wrap_PI_tests[i].wv) > 0.001f) {
             hal.console->printf("wrap_PI: v=%f wv=%f r=%f\n",
                                 wrap_PI_tests[i].v,
                                 wrap_PI_tests[i].wv,
